Remove ExpiringTokenBasedRememberMeServices, add ExpiringTokenBasedRememberMeHandler
